거너 스킬 오류 수정
저격 모션 중 죽고 부활해서 다시 저격쐈을 때 저격이 안풀리는 버그 수정
수류탄 데미지 안들어가는 버그 수정

+++ Assets/Scripts/Skills/GrenadeChildScript.cs
@@ -33,7 +33,7 @@ public class GrenadeChildScript : MonoBehaviour
             if (targetPhotonView != null)
             {
                 // 몬스터의 체력 감소 처리
-                targetPhotonView.RPC("TakeDamage", RpcTarget.All, new object[] { damage, parentGrenadeScript.attackerViewId, isCriticalHit, transform.position, 1f });
+                targetPhotonView.RPC("TakeDamage", RpcTarget.All, new object[] { (int)damage, parentGrenadeScript.attackerViewId, isCriticalHit, transform.position, 1f });
             }
         }
         if (col.CompareTag("Removable Obstacle"))
